refactor: improve role description and update package installation
The commit updates the description of the Ansible role to be more concise and removes unnecessary details. It also refactors the tasks file by adding a header comment and removing redundant tasks related to gathering facts. Additionally, it updates the package installation task by removing the fact gathering task for distribution information on Debian-based systems.